removing bug from depth texture, now the texture is screen size unlike before and depth data also is normalized by linear interpolation

+++ src/realsenseimage.h
@@ -61,7 +61,7 @@ protected:
   void updateVertices(const std::vector<float> &vertsAndTex,
                       const std::vector<unsigned> &indices);
 
-  std::vector<unsigned> _rgb;
+  std::vector<uint8_t>  _rgb;
 
   std::vector<float> _buffer;
   unsigned _width, _height;
